refactor: move options objects to interfaces
