+++ src/interfaces/client/constructorOptions.ts
@@ -1,0 +1,6 @@
+export interface IConstructorOptions {
+  baseUrl: string,
+  apiKey: string,
+  userAgent: string,
+  params?: Record<string, string>,
+}
