+++ src/interfaces/client/extensionOptions.ts
@@ -1,0 +1,6 @@
+export interface IExtendOptions {
+  baseUrl?: string,
+  apiKey?: string,
+  userAgent?: string,
+  params?: Record<string, string>,
+}
