+++ src/interfaces/client/index.ts
@@ -1,0 +1,1 @@
+export * from './constructorOptions';
