+++ src/interfaces/client/requestOptions.ts
@@ -1,0 +1,9 @@
+import { Method } from 'axios';
+
+export interface IRequestOptions {
+  path: string,
+  paginated?: boolean,
+  method?: Method,
+  params?: Record<string, any>,
+  json?: Record<string, string>,
+}
