+++ src/interfaces/mixins/index.ts
@@ -1,3 +1,2 @@
-export * from './module';
 export * from './managerMixinConstructor';
 export * from './resourceMixinConstructor';
