+++ src/interfaces/mixins/managerMixinConstructor.ts
