+++ src/interfaces/mixins/resourceMixinConstructor.ts
