+++ src/interfaces/paginator/paginationOptions.ts
@@ -1,0 +1,7 @@
+import { AxiosInstance } from 'axios';
+
+export interface IPaginationOptions {
+  client: AxiosInstance,
+  path: string,
+  params?: Record<string, string>
+}
